refactor(watcher): split findings.py + _util.py out of agent.py
agent.py dropped from 2023 → 1450 LOC by moving the finding model, dedup,
persistence, status lifecycle, formatting, compaction, and KG escalation
into agents/watcher/findings.py (587 LOC). A new agents/watcher/_util.py
(89 LOC) holds log/path/hash leaf helpers so findings.py can depend on
them without circular imports.

Identity, scanning, and CLI orchestration stay in agent.py. surface_pending
and _post_resolution_event also stay — they couple findings logic to the
identity block (_do_checkin / get_watcher_identity).

No behavioral changes. 134 watcher tests + full 6770-test suite all pass.
Test fixture updated: _isolate_watcher_state now patches the split-out
modules at source, and three helper patchers (_mock_post_finding,
_mock_escalate_to_kg, _mock_watcher_identity) mirror patches across the
importlib watcher_module copy, the real agents.watcher.agent module, and
findings module so the test monkeypatches reach every call site.
